requirements.txt
asgiref==3.4.1
attrs==21.4.0
blis==0.7.8
catalogue==2.0.8
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.6
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.7.0
django-filter==21.1
django-guardian==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
en-core-web-md==3.4.0
fr-core-news-md==3.4.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.9.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
langcodes==3.3.0
Markdown==3.3.6
MarkupSafe==2.1.0
murmurhash==1.0.8
numpy==1.23.2
packaging==21.3
pathy==0.6.2
preshed==3.0.7
psycopg2-binary==2.9.1
pydantic==1.9.2
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.2
srsly==2.4.4
swagger-spec-validator==2.7.4
thinc==8.1.0
tqdm==4.64.0
typer==0.4.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.8
wasabi==0.10.1
whitenoise==5.3.0
zipp==3.6.0
